refactor: lift blob decoders into sracha-vdb::blob_codecs
Pure move, no behavior change. `decode_raw`, `decode_zip_encoding`,
`decode_quality_encoding`, `decode_irzip_column`, and
`expand_via_page_map` now live in `sracha-vdb` so both the fastq
pipeline and the upcoming `sracha vdb dump` command share one codec
surface. `sracha-core::pipeline::blob_decode` re-exports them under
the same names. Error::Pipeline → sracha_vdb::Error::Format inside
the moved bodies; call sites that ? through `sracha-core::Error`
pick up the `From<sracha_vdb::Error>` conversion transparently.
